+++ FolderDiffIL4DotNet.Tests/AppDataOverrideScope.cs
@@ -47,6 +47,9 @@ namespace FolderDiffIL4DotNet.Tests.Helpers
         internal string UserConfigFileAbsolutePath
             => Path.Combine(ApplicationDataRootAbsolutePath, "config.json");
 
+        internal string HtmlReportChecklistFileAbsolutePath
+            => Path.Combine(ApplicationDataRootAbsolutePath, "HtmlReport", "checklist.json");
+
         public void Dispose()
         {
             if (_disposed)